requirements.txt
# sorl-thumbnail импортирует pkg_resources; в setuptools 81+ модуль убран
setuptools>=65,<81
Django==3.2.10
django-allauth==0.51.0
django-ckeditor==6.3.2
django-cleanup==6.0.0
django-recaptcha3==0.4.0
djangorestframework==3.14.0
idna==3.4
idna-ssl==1.1.0
lxml==4.9.3
# mysqlclient==2.1.1
openpyxl==3.1.2
Pillow==8.4.0
pytils==0.3
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.8.0
telepot==12.7
transliterate==1.10.2
typing_extensions
urllib3
whitenoise==6.5.0
wrapt==1.15.0
yarl==1.9.2
yookassa==2.4.0
beautifulsoup4==4.12.3
selenium==4.25.0
webdriver-manager==4.0.2
wheel==0.37.0
